Move `concentration_measurement` into `measurement`
